# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Omreca Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
        "/auth/create": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auths"
                ],
                "parameters": [
                    {
                        "description": "Create Auth Json",
                        "name": "Auth_create",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_domain.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth_domain.LoginResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auths"
                ],
                "parameters": [
                    {
                        "description": "Auth JSON",
                        "name": "Auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth_domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth_domain.LoginResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth_domain.CreateUserRequest": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "integer"
                },
                "user_password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth_domain.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth_domain.LoginResponse": {
            "type": "object",
            "properties": {
                "createdby": {
                    "type": "string"
                },
                "createdt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isactive": {
                    "type": "integer"
                },
                "isadmin": {
                    "type": "integer"
                },
                "updatedby": {
                    "type": "string"
                },
                "updatedt": {
                    "type": "string"
                },
                "user_password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}